Refactored transmission details
